+++ lib/folio/engine.rb
@@ -99,6 +99,15 @@ module Folio
     config.folio_users_include_nickname = true
     config.folio_users_confirmable = false
     config.folio_users_confirm_email_change = true
+    config.folio_users_device_modules = %i[
+      database_authenticatable
+      recoverable
+      rememberable
+      trackable
+      invitable
+      timeoutable
+      lockable
+    ]
     config.folio_users_publicly_invitable = false
     config.folio_users_use_address = true
     config.folio_users_omniauth_providers = %i[facebook google_oauth2 twitter2 apple]